[C] Use size_t for queue capacity and indexing.

+++ aeron-client/src/main/c/concurrent/aeron_mpsc_concurrent_array_queue.c
@@ -17,11 +17,11 @@
 #include "aeron_alloc.h"
 #include "concurrent/aeron_mpsc_concurrent_array_queue.h"
 
-int aeron_mpsc_concurrent_array_queue_init(aeron_mpsc_concurrent_array_queue_t *queue, uint64_t length)
+int aeron_mpsc_concurrent_array_queue_init(aeron_mpsc_concurrent_array_queue_t *queue, size_t length)
 {
-    length = (uint64_t)aeron_find_next_power_of_two((int32_t)length);
+    length = (size_t)aeron_find_next_power_of_two((int32_t)length);
 
-    if (aeron_alloc((void **)&queue->buffer, sizeof(void *) * (size_t)length) < 0)
+    if (aeron_alloc((void **)&queue->buffer, sizeof(void *) * length) < 0)
     {
         return -1;
     }
@@ -50,10 +50,10 @@ int aeron_mpsc_concurrent_array_queue_close(aeron_mpsc_concurrent_array_queue_t 
 extern aeron_queue_offer_result_t aeron_mpsc_concurrent_array_queue_offer(
     aeron_mpsc_concurrent_array_queue_t *queue, void *element);
 
-extern uint64_t aeron_mpsc_concurrent_array_queue_drain(
-    aeron_mpsc_concurrent_array_queue_t *queue, aeron_queue_drain_func_t func, void *clientd, uint64_t limit);
+extern size_t aeron_mpsc_concurrent_array_queue_drain(
+    aeron_mpsc_concurrent_array_queue_t *queue, aeron_queue_drain_func_t func, void *clientd, size_t limit);
 
-extern uint64_t aeron_mpsc_concurrent_array_queue_drain_all(
+extern size_t aeron_mpsc_concurrent_array_queue_drain_all(
     aeron_mpsc_concurrent_array_queue_t *queue, aeron_queue_drain_func_t func, void *clientd);
 
-extern uint64_t aeron_mpsc_concurrent_array_queue_size(aeron_mpsc_concurrent_array_queue_t *queue);
+extern size_t aeron_mpsc_concurrent_array_queue_size(aeron_mpsc_concurrent_array_queue_t *queue);
